// src/app/api/model/reclamationDto.ts
/**
 * GestionRFQ API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ReclamationType } from './reclamationType';
import { ReclamationStatus } from './reclamationStatus';


export interface ReclamationDto { 
    id?: number;
    title?: string | null;
    message?: string | null;
    type?: ReclamationType;
    status?: ReclamationStatus;
    createdAt?: string;
    updatedAt?: string | null;
    userId?: number | null;
}
export namespace ReclamationDto {
}