// src/app/api/model/reclamationStatus.ts
/**
 * GestionRFQ API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type ReclamationStatus = 0 | 1 | 2 | 3;

export const ReclamationStatus = {
    NUMBER_0: 0 as ReclamationStatus,
    NUMBER_1: 1 as ReclamationStatus,
    NUMBER_2: 2 as ReclamationStatus,
    NUMBER_3: 3 as ReclamationStatus
};